+++ stories/axis.tsx
@@ -479,7 +479,7 @@ storiesOf('Axis', module)
       const newArray: (string | number)[] = [];
       stringArray.forEach((element: string) => {
         // eslint-disable-next-line no-unused-expressions
-        !isNaN(parseInt(element)) ? newArray.push(parseInt(element)) : newArray.push(element);
+        !isNaN(parseFloat(element)) ? newArray.push(parseFloat(element)) : newArray.push(element);
       });
       return newArray;
     }
